refactor: dto 를 model 패키지에서 분리해 view 가 model 를 참조하지 않도록 해라
