+++ src/main/java/lotto/controller/LottoController.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.function.Supplier;
 import lotto.handler.InputHandler;
 import lotto.model.customer.Customer;
-import lotto.model.dto.LottoDto;
-import lotto.model.dto.ResultDto;
+import lotto.dto.LottoDto;
+import lotto.dto.ResultDto;
 import lotto.model.lotto.Lotto;
 import lotto.model.lotto.WinningLotto;
 import lotto.service.LottoResultService;
