+++ src/main/java/lotto/dto/LottoDto.java
@@ -1,4 +1,4 @@
-package lotto.model.dto;
+package lotto.dto;
 
 import java.util.List;
