+++ src/main/java/lotto/dto/ResultDto.java
@@ -1,4 +1,4 @@
-package lotto.model.dto;
+package lotto.dto;
 
 import java.util.ArrayList;
 import java.util.List;
